Refactors subscription form: now it is just another section
