+++ public/locales/be/common.json
@@ -4,7 +4,9 @@
   "ru": "RU",
   "title": "Мара",
   "description": "суполка беларусаў у Нідэрландах",
-  "subscribe-title": "Мы распрацоўваем партал нашай арганізацыі. Падпісвайцеся, каб быць у курсе",
+  "news-title": "Навіны",
+  "subscribe-title": "Падпісвайцеся",
+  "subscribe-text": "Мы гарантуем ананімнасьць. Ваша пошта можа быць бачна толькі разработчыкам",
   "subscribe-input-placeholder": "email",
   "subscribe-button": "Падпісацца",
   "subscribed-text": "Дзякуй за вашу зацікаўленнасць!"
