+++ public/locales/ru/common.json
@@ -4,7 +4,9 @@
   "ru": "RU",
   "title": "Мара",
   "description": "сообщество беларусов в Нидерландах",
-  "subscribe-title": "Мы разрабатываем портал нашей организации. Подписывайтесь, чтобы быть в курсе",
+  "news-title": "Новости",
+  "subscribe-title": "Подписывайтесь",
+  "subscribe-text": "Мы гарантируем анонимность. Ваша почта может быть видна только разработчикам",
   "subscribe-input-placeholder": "email",
   "subscribe-button": "Подписаться",
   "subscribed-text": "Спасибо за вашу заинтересованность!"
